application/requirements.txt
requests
boto3==1.34.95
botocore==1.34.95
psycopg2-binary==2.9.9
